# api/openapi.yaml
openapi: 3.0.3
info:
  title: Poydem API
  version: 1.0.0
  license:
    name: Proprietary
  description: |
    Final frontend/backend contract for Poydem MVP API v1.

    All enum values use lower-case except stable error and system-reason codes,
    which use UPPER_SNAKE_CASE. Every JSON response uses UTF-8. Every response
    includes X-Request-ID. Timestamps are RFC 3339 date-time values with an offset.

    Access tokens are sent as Bearer JWTs. Refresh tokens exist only in an HttpOnly
    cookie and are never returned in JSON or URLs. Lists documented with pagination
    return `items: []` and `totalPages: 0` when empty. Unexpected server failures
    return HTTP 500 with error.code `INTERNAL_ERROR` in the common error envelope.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /users/me/avatar:
    post:
      operationId: uploadMyAvatar
      tags: [Users]
      summary: Upload avatar
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED]
        '413': [AVATAR_TOO_LARGE]
        '415': [UNSUPPORTED_AVATAR_TYPE]
        '500': [AVATAR_UPLOAD_FAILED]
        '503': [AVATAR_UPLOAD_FAILED]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              additionalProperties: false
              properties: {file: {type: string, format: binary}}
      responses:
        '200':
          description: Uploaded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AvatarResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    delete:
      operationId: deleteMyAvatar
      tags: [Users]
      summary: Delete avatar
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED]
      responses:
        '204': {description: Deleted}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [Admin]
      summary: Admin dashboard
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, FORBIDDEN]
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminDashboard'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [Admin]
      summary: Admin users
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, FORBIDDEN]
      parameters:
        - {name: search, in: query, schema: {type: string}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [Admin]
      summary: Admin user details
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, FORBIDDEN]
        '404': [USER_NOT_FOUND]
      parameters: [{$ref: '#/components/parameters/UserId'}]
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly cookie managed by the browser; JavaScript cannot read it
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    BadRequest:
      description: Bad request
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Unauthorized:
      description: Unauthorized
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: Forbidden
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: Not found
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    PayloadTooLarge:
      description: Uploaded avatar exceeds 5 242 880 bytes
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    UnsupportedMediaType:
      description: Avatar must be image/jpeg, image/png or image/webp
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    InternalError:
      description: Internal server error
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ServiceUnavailable:
      description: Temporarily unavailable
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  schemas:
    DictionaryItem:
      type: object
      required: [id, name, slug]
      additionalProperties: false
      properties:
        id: {type: integer, format: int64, minimum: 1}
        name: {type: string, minLength: 1}
        slug: {type: string, nullable: true}
    UserShort:
      type: object
      required: [id, firstName, lastName, avatarUrl]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        firstName: {type: string}
        lastName: {type: string, nullable: true}
        avatarUrl: {type: string, format: uri, nullable: true}
    User:
      allOf:
        - {$ref: '#/components/schemas/UserShort'}
        - type: object
          required: [city, about, interests, role, status, isProfileComplete, createdAt]
          properties:
            city:
              type: object
              nullable: true
              allOf: [{$ref: '#/components/schemas/DictionaryItem'}]
            about: {type: string, nullable: true}
            interests: {type: array, items: {$ref: '#/components/schemas/DictionaryItem'}}
            role: {$ref: '#/components/schemas/UserRole'}
            status: {$ref: '#/components/schemas/UserStatus'}
            isProfileComplete:
              type: boolean
              description: Calculated by backend as non-empty firstName and non-null city
            createdAt: {type: string, format: date-time}
    UserRole:
      type: string
      enum: [user, admin]
    UserStatus:
      type: string
      enum: [active, banned]
...
    UserList:
      type: object
      required: [items, pagination]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/User'}}
        pagination: {$ref: '#/components/schemas/Pagination'}
...
    AdminDashboard:
      type: object
      required: [usersTotal, activeEvents, pendingReports, newRegistrations30d]
      additionalProperties: false
      properties:
        usersTotal: {type: integer, minimum: 0}
        activeEvents: {type: integer, minimum: 0}
        pendingReports: {type: integer, minimum: 0}
        newRegistrations30d: {type: integer, minimum: 0, description: Registrations during the last 30 calendar days including today}
...
    AvatarResponse:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl: {type: string, format: uri}
...
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: {type: integer, minimum: 1}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
        totalPages: {type: integer, minimum: 0, description: Zero when total is zero}
    ErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, details]
          additionalProperties: false
          properties:
            code: {$ref: '#/components/schemas/ErrorCode'}
            message: {type: string, example: No free places in company}
            details: {$ref: '#/components/schemas/ErrorDetails'}
    ErrorDetails:
      type: object
      properties:
        fields:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
          description: Present for VALIDATION_ERROR; keys are request field or parameter names
      additionalProperties: true
    ErrorCode:
      type: string
      description: Stable machine-readable code. Frontend logic must never branch on error.message.
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - INTERNAL_ERROR
        - OAUTH_PROVIDER_UNAVAILABLE
        - USER_BANNED
        - USER_NOT_FOUND
        - PROFILE_INCOMPLETE
        - EVENT_NOT_FOUND
        - EVENT_NOT_AVAILABLE
        - INVALID_STATUS_TRANSITION
        - COMPANY_NOT_FOUND
        - COMPANY_FULL
        - COMPANY_CLOSED
        - COMPANY_BLOCKED
        - COMPANY_CAPACITY_BELOW_MEMBERS
        - NOT_COMPANY_OWNER
        - NOT_COMPANY_MEMBER
        - OWNER_CANNOT_LEAVE
        - OWNER_CANNOT_BE_REMOVED
        - ALREADY_COMPANY_MEMBER
        - ALREADY_IN_EVENT_COMPANY
        - APPLICATION_ALREADY_EXISTS
        - APPLICATION_NOT_FOUND
        - APPLICATION_ALREADY_RESOLVED
        - ALREADY_EVENT_PARTICIPANT
        - NOT_SOLO_PARTICIPANT
        - REPORT_TARGET_NOT_FOUND
        - REPORT_NOT_FOUND
        - REPORT_ALREADY_RESOLVED
        - AVATAR_TOO_LARGE
        - UNSUPPORTED_AVATAR_TYPE
        - AVATAR_UPLOAD_FAILED
        - EVENT_FORBIDDEN
        - EVENT_NOT_EDITABLE
        - EVENT_COVER_TOO_LARGE
        - UNSUPPORTED_EVENT_COVER_TYPE
        - EVENT_COVER_UPLOAD_FAILED
        - RATE_LIMITED
        - APPLICANT_BANNED
        - CANNOT_CHANGE_OWN_STATUS
        - CANNOT_BAN_ADMIN